requirements.txt
    ```text
    torch>=2.0.0
    torchvision
    numpy
    scipy
    matplotlib
    Pillow
    scikit-image
    ```